Resolución de fallos críticos en la 2.18: Si el canal no está compilado previamente y la actualización automática está activada, falla al entrar

+++ plugins/video/plugin.video.pelisalacarta/resources/lib/downloadtools.py
@@ -895,6 +895,7 @@ def updatechannel(channelName):
 		outfile.close()
 		logger.info("Grabado a " + channelFileSource)
 		
-		os.remove(channelFileCompiled)
+		if os.path.exists(channelFileCompiled):
+			os.remove(channelFileCompiled)
 
 	return updated
